BulkLoad coalescing: address review — int64 getBytes() + guard unknown-size sentinel
Review of the coalescing change flagged that BulkLoadJobFileManifestEntry::getBytes()
returned size_t while the byte-count field is set from BulkLoadManifest::getTotalBytes()
(whose default is the -1 'unknown' sentinel). A -1 stored into the size_t field becomes
SIZE_MAX and, cast back to int64_t at the call sites, yields -1 -- silently DECREMENTing
the coalescing byte sums. Return int64_t and clamp unknown/unset to 0 so an unknown-size
manifest contributes nothing to the byte target (the task then falls to the
MANIFEST_COUNT_MAX_PER_BULKLOAD_TASK cap, as intended) and drop the now-redundant
static_cast at the two call sites.

Also note in the DD_BULKLOAD_TASK_TARGET_RATIO knob comment that
MANIFEST_COUNT_MAX_PER_BULKLOAD_TASK caps manifests/task, so for dumps whose manifests are
smaller than ~target/cap the coalesced shard can still land below DD's merge floor.

+++ fdbserver/datadistributor/DataDistribution.cpp
@@ -2051,7 +2051,7 @@ Future<Void> scheduleBulkLoadJob(Reference<DataDistributor> self, Promise<Void> 
 	if (SERVER_KNOBS->DD_BULKLOAD_TASK_TARGET_RATIO > 0) {
 		int64_t datasetBytes = 0;
 		for (const auto& manifestEntryPair : *self->bulkLoadJobManager.get().manifestEntryMap) {
-			datasetBytes += static_cast<int64_t>(manifestEntryPair.second.getBytes());
+			datasetBytes += manifestEntryPair.second.getBytes();
 		}
 		bulkLoadTaskTargetBytes = static_cast<int64_t>(getMaxShardSize(static_cast<double>(datasetBytes)) *
 		                                               SERVER_KNOBS->DD_BULKLOAD_TASK_TARGET_RATIO);
@@ -2136,7 +2136,7 @@ Future<Void> scheduleBulkLoadJob(Reference<DataDistributor> self, Promise<Void> 
 						ASSERT(it != self->bulkLoadJobManager.get().manifestEntryMap->end());
 						manifestEntry = it->second;
 						manifestEntries.push_back(manifestEntry);
-						accumulatedBytes += static_cast<int64_t>(manifestEntry.getBytes());
+						accumulatedBytes += manifestEntry.getBytes();
 						beginKey = manifestEntry.getEndKey();
 						if (bulkLoadTaskTargetBytes > 0 && accumulatedBytes >= bulkLoadTaskTargetBytes) {
 							break; // reached ~target shard size; cut the task here
